+++ exerc01/summary-parallel.py
@@ -6,7 +6,7 @@ def do_plot(csv_filename):
     summary = pd.read_csv(csv_filename)
 
     # the regular hmm-generated:
-    if csv_filename == 'summary.csv':
+    if csv_filename == 'summary.csv' or csv_filename == 'hmm-summary.csv':
         selection = summary[['N', 'M', 'I', 'Test avgAc']]
 
     # ad hoc for other tests:
@@ -17,7 +17,7 @@ def do_plot(csv_filename):
         selection = summary[['M', 'Train avgAc', 'Test avgAc']]
 
     else:
-        print('unrecognized file: {}'.format(csv_filename))
+        print(f'unrecognized file: {csv_filename}')
         return
 
     fig = px.parallel_coordinates(selection,